MODCLUSTER-497 MODCLUSTER-498 Tomcat modeler fixes and code cleanup: * MODCLUSTER-498 Extract Tomcat modeler specific methods from ModClusterConfig into separate Tomcat-specific class * MODCLUSTER-497 ModClusterConfig#setAdvertiseInterface(java.lang.String) should not have been deprecated as it's used by Tomcat modeller
